tools/requirements.txt
influxdb-client==1.48.0
pymcuprog==3.17.3.45
pyserial==3.5
